% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mfv}
\alias{mfv1}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{statip}{\code{\link[statip:mfv]{mfv()}}, \code{\link[statip:mfv1]{mfv1()}}}
}}

